Fix IssueInstant date format
I'm having problems trying to authenticate using ruby-saml, and after some investigation I've discovered that the timestamp format is incorrect for my IdP.

After some searching and looking for other SAML examples, I've start to believe that the timestamp format without the trailing 'Z' was in fact incorrect. So I'm proposing to fix this issue inside the gem.

There are something that I missing?