Introduce sp_entity_id and use it to #validate_audience
The #validate_audience method matched the audience in the
SAML response with the settings#issuer, which is equal to
the IDP entity id. The audience should match the SP,
rather than the IDP.

This commit introduces #sp_entity_id as an option on settings and
has #validate_audience validate against it.

Closes #283