Test for self closed audience tag
After discussion on:
https://github.com/onelogin/ruby-saml/issues/443

We determined that it was related to the way that the text is being
returned and checked. In my previous commit I applied the change
discussed in the issue above.

This commit adds a new response and test that fails before the and
passes once the change is added.

I used a SAML tool to produce a new response that has the unexpected
selfclosing tag and included it in the test/responses folder.