Support decimal seconds in cacheDuration parsing
Since the introduction of cacheDuration parsing in 25cbddd we've been
seeing parsing failures for one of our IdPs, whose cacheDuration value
is set to

    cacheDuration="PT6H0M0.000S"

This is a valid ISO8601 duration, however the regexp being used for
parsing doesn't provide for the possibility of milliseconds.

The ISO8601 spec[1] supports decimal values for at least some of a
duration's components (S4.4.3.2):

  > If necessary for a particular application, the lowest order
  > components may have a decimal fraction.

It's not entirely plain which components this might refer to, however
the xsd:duration spec[2] is plainer:

  > The numbers may be any unsigned integer, with the exception of the
  > number of seconds, which may be an unsigned decimal number.

This commit therefore adds fractional seconds support to the duration
parsing. To ensure that we actually respect these values, I've updated
our parsed value coercion to produce floats for all components. This is
arguably unnecessary for the non-float components (since the regexp will
only accept integers), but it allows us to reduce the code with no ill
effects.

The ISO spec permits arbitrary decimal precision, and supports both
commas and dots as the decimal separator, so since the xsd:duration
schema doesn't mention either concern, I've followed the ISO spec here.

[1] https://www.loc.gov/standards/datetime/iso-tc154-wg5_n0038_iso_wd_8601-1_2016-02-16.pdf
[2] http://www.datypic.com/sc/xsd/t-xsd_duration.html