Add unit coverage for Utils.parse_duration
This adds some more comprehensive unit testing for the .parse_duration
helper, since we'd like to modify it for better coverage of the ISO8601
duration spec.

Since the helper both parses the duration and adds it to the current
time (by default) before returning it, I've used the unix epoch as a
reference point instead and formatted the result as an ISO timestamp
before asserting, to hopefully make any errors clearer.

I've not gone for full coverage, but have tried to test the main edge
cases. In subsequent commits I'll refactor the parsing and introduce
support for fractional values.