Simplify dependency constraints in gemspec
Looking at the latest release, 1.16.0 it depends on the
it depends on everything in the `else` branches in the gemspec.
https://rubygems.org/gems/ruby-saml/versions/1.16.0

Runtime

    nokogiri >= 1.13.10
    rexml >= 0

Development

    minitest ~> 5.5, < 5.19.0
    mocha ~> 0.14
    pry-byebug >= 0
    rake >= 12.3.3
    shoulda ~> 2.11
    simplecov < 0.22.0
    simplecov-lcov > 0.7.0
    systemu ~> 2
    timecop ~> 0.9

There has also never been a JRuby specific version published:
https://rubygems.org/gems/ruby-saml/versions
so remove all special handling for JRuby.

If CI shows any need for it, we can add back constraints in the Gemfile.