Implement correct handling of CBC padding for symmetric decryption.
AES and 3DES, when used with encrypted assertions with CBC padding, use a form of padding called ISO 10126 and is documented here: https://www.w3.org/TR/xmlenc-core1/#sec-Padding.
The current implementation calls this "noise" and uses a regular expression to attempt to extract the plaintext without the padding. Rather than use a regular expression, this implements correct handling of the padding.